backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
cohere==4.37
python-dotenv==1.0.0
